device: add LiveGetDeviceRunmode api
In some functions, such as DetectEngineMultiTenantSetupLoadLivedevMappings
and EBPFLoadFile, LiveGetDevice is called to
search for a device, but it will need the runmode id as parameter.
It's needed to add an api that given the device name it returns the runmode id
because at this stage the device's runmode id is not knew.