init: change uid/gid on system without libcap-ng
capng_change_id() is based on setresuid/gid() to change
the process uid/gid according to the 'run-as' settings.

This changeset uses directly setresuid/gid() which noops
when built with libcap-ng support.

On system without capabilities, the uid/gid change is done later in the init
phase to allow the worker threads to open their devices before the priv drop.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@cognitix.de>