reputation: add ip from unix socket
This introduces some changes in such a way as to be able
to update IP reputation adding IP addresses through unix-socket.

All the IP addresses sent are stored in a list,
and when the rules are reloaded iprep loads the IPs from the list.

It's possible to have at most 100 IPs in pending,
then the user is forced to run 'reload-rules' to be able to add them
again until the maximum number of entries via unix-socket is reached,
which is 500 IPs.

Ticket #659.