privs: add function to raise or drop priv levels.
This function set allows the source thread to temporarily regain the privileges
the process previously had to perform specific high permission task.
Once done, the privileges must be dropped again.

As the effective uid/gid is set per-process, those function are coupled with a
mutex to serialize the raise and drop of privileges.

Signed-off-by: Emmanuel Roullit <emmanuel.roullit@cognitix.de>