Always store the vlan id of packets
If the setting vlan.use-for-tracking was set to false, suricata would
in some cases ignore the vlan information of packets and leave them 0
for FlowHash calculations and Flow comparisons. This commit instead
always fills in the vlan_id field, but passes 0 for the vlan_id in the
FlowHash calculation and Flow comparison. This is done using a bitmask
to avoid introducing additional branches, as suggested by Victor Julien
in the suricata IRC.

Also turned FLOW_CMP from a macro into an inline function at Victor's
request.