protodetect: retry when both sides do not agree
This fixes a scenario with an evil server (and a legit client)
The evil server may try to first speak another protocol
So, Suricata will end classifying the flow with this faked protocol
as in case of protocol mismatch between client and server, the
chosen protcol is the one spoken by the server.

This patch fixes this by retrying protocol detection in the case
of a protocol mismatch between client and server.