detect/dce-iface: Fix check for version in regex
The code considered that for a return value of 5 from `pcre_exec`, the
argument shall have the `any_frag` option and version. This was not
taking into consideration the case where argument string can be
`12345678-1234-1234-1234-123456789ABC,any_frag`. This worked out well so
far as the version in this case was an empty string converted to 0 by
atoi. Now that atoi occurences have been removed, make the check
stricter by checking the offset vector's value for version alongwith the
return value.