configure: change version logic
Consider all builds from a non-git directory 'release'.

If in a git checkout see if the version from autoconf (PACKAGE_VERSION)
is part of a git tag. If it is, and the current branches hash matches
the tag, then consider it release. If not, consider 'dev' version.

This way we can see if a build is RELEASE of a git dev version in the
output easily, w/o needing to specifically tag the release as such like
is done now.