WIP decode/store: generic storage for decode layer
A fixed size static store is part of the Packet structure. If that space
isn't sufficient the store can dynamically expand. It uses a list of
storage blocks for this.

The primary goals are:
- don't use space in the packet when a feature/proto isn't in use
- reduce (or avoid) the need for new protocols to manually add to the
  Packet structure

In general the storage is a TLV (type length value) store, but some types
get special support to reduce the space overhead of the TLV header.

Implemented so far:
- VLAN (2 bytes)
- MPLS (4 bytes)
- TLV (3 bytes + the data to be stored)