output-json: introduce stream data logging
This patch adds a stream field for stream based protocol and allow
non streamed based protocol to add a payload and/or payload_printable
field(s) for non streamed based protocols.

In the case of stream based protocols, we display a stream.client
and a stream.server object that contain payload and/or
payload_printable fields.