flow: clean expectations at flow end
When a flow timeout, we can have still existing expectations that
are linked to this flow. Given that there is a delay between the
real ending of the flow and its destruction by Suricata, we can
assume the risk to clean the expectations that have been triggered
by the to-be-deleted flow.