streaming/api: fix overlap check
In some cases a SBB could be seen as overlapping with the requested
offset, when it was in fact precisely before it. In some special cases
this could lead to the stream engine not progressing the 'raw' progress.

(cherry picked from commit 2c050187a3ad78388e57781c4feb74426bc75e21)