rust distcheck fixup: do not distribute Cargo.lock
Stop including Cargo.lock in the distribution archive
as this causes issues in out of tree builds.

Instead generate it as needed, which is already done
for builds where the Rust dependencies are not
vendored.

And when they are vendored, we can use the vendored
sources to regenerate Cargo.lock without going
online.

With Rust 1.36 we can ensure we never go online with
cargo --offline, but that flag is not available in
versions older than 1.36.