python: add --with-target-python option
When installing python scripts, distutils would use the python used to
run setup.py as shebang for the scripts it installs.

However, when cross-compiling, this is most often not correct.

Instead, add a --with-target-python option to allow the user to set its
own value such as '/usr/bin/env python'. To keep current behavior,
$TARGET_PYTHON is set to $HAVE_PYTHON by default

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>